Consolidate legacy dialect files; drop test data from production package
