+++ pkg/builderapi/legacy/register_validators.go
@@ -9,7 +9,10 @@ import (
 	"net/http"
 
 	apiv1 "github.com/ethpandaops/go-eth2-client/api/v1"
+	"github.com/ethpandaops/go-eth2-client/spec/phase0"
 	"github.com/sirupsen/logrus"
+
+	"github.com/ethpandaops/buildoor/pkg/signer"
 )
 
 // HandleRegisterValidators handles POST /eth/v1/builder/validators.
@@ -84,4 +87,61 @@ func (h *Handler) HandleRegisterValidators(w http.ResponseWriter, r *http.Reques
 
 	log.WithField("stored_count", len(regs)).Info("Validator registrations accepted")
 	w.WriteHeader(http.StatusOK)
+}
+
+// VerifyRegistration verifies the BLS signature of a validator registration
+// using DOMAIN_APPLICATION_BUILDER with zero parameters (for tests).
+// For chain-specific verification (e.g. mev-boost registrations), use VerifyRegistrationWithDomain.
+func VerifyRegistration(reg *apiv1.SignedValidatorRegistration) bool {
+	var zero phase0.Version
+	var zeroRoot phase0.Root
+	return VerifyRegistrationWithDomain(reg, zero, zero, zeroRoot)
+}
+
+// VerifyRegistrationWithDomain verifies the BLS signature of a validator registration
+// using DOMAIN_APPLICATION_BUILDER. Tries (0,0), (genesisForkVersion, 0) from the beacon
+// (mev-boost-relay style), then (forkVersion, genesisValidatorsRoot). Genesis fork version
+// is taken from the beacon so local devnets with their own fork versions work.
+func VerifyRegistrationWithDomain(reg *apiv1.SignedValidatorRegistration, genesisForkVersion, forkVersion phase0.Version, genesisValidatorsRoot phase0.Root) bool {
+	if reg == nil || reg.Message == nil {
+		return false
+	}
+
+	messageRoot, err := reg.Message.HashTreeRoot()
+	if err != nil {
+		return false
+	}
+
+	var root phase0.Root
+	copy(root[:], messageRoot[:])
+
+	var zeroVersion phase0.Version
+	var zeroRoot phase0.Root
+
+	// 1) (0, 0) — some clients use this for DOMAIN_APPLICATION_BUILDER.
+	domainZero := signer.ComputeDomain(signer.DomainApplicationBuilder, zeroVersion, zeroRoot)
+	signingRootZero := signer.ComputeSigningRoot(root, domainZero)
+	if signer.VerifyBLSSignature(reg.Message.Pubkey, signingRootZero[:], reg.Signature) {
+		return true
+	}
+
+	// 2) (genesisForkVersion, 0) — mev-boost-relay style; genesis fork from beacon (devnet-friendly).
+	if genesisForkVersion != zeroVersion {
+		domainRelay := signer.ComputeDomain(signer.DomainApplicationBuilder, genesisForkVersion, zeroRoot)
+		signingRootRelay := signer.ComputeSigningRoot(root, domainRelay)
+		if signer.VerifyBLSSignature(reg.Message.Pubkey, signingRootRelay[:], reg.Signature) {
+			return true
+		}
+	}
+
+	// 3) (forkVersion, genesisValidatorsRoot) — chain-specific domain.
+	if forkVersion != zeroVersion || genesisValidatorsRoot != zeroRoot {
+		domainChain := signer.ComputeDomain(signer.DomainApplicationBuilder, forkVersion, genesisValidatorsRoot)
+		signingRootChain := signer.ComputeSigningRoot(root, domainChain)
+		if signer.VerifyBLSSignature(reg.Message.Pubkey, signingRootChain[:], reg.Signature) {
+			return true
+		}
+	}
+
+	return false
 }
